MP-8785: Added the following fields to AllNoTypes entity:

+++ src/Nox.Types.Abstractions/Enums/NoxType.cs
@@ -128,10 +128,10 @@ public enum NoxType : uint
     [SimpleType(typeof(string))]
     DateTimeSchedule = 3271762546,
 
-    [SimpleType(typeof(byte))]
+    [SimpleType(typeof(ushort))]
     DayOfWeek = 491349974,
 
-    [SimpleType(typeof(double))]
+    [SimpleType(typeof(decimal))]
     Distance = 1460893529,
 
     [SimpleType(typeof(string))]
